requirements.txt
arrow==1.2.3
attrs==22.2.0
Babel==2.11.0
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
flask-dance==6.2.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
greenlet==2.0.2
ics==0.7.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
nltk==3.8.1
oauthlib==3.2.2
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.3
TatSu==4.4.0
tqdm==4.65.0
typing-extensions==4.4.0
urllib3==1.26.14
URLObject==2.4.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.13.0